split existing test cases into multiple files (just moved code around, no further changes)
